requirements.txt
torch>=1.13.0
torchvision>=0.14.0
timm>=0.9.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.64.0
opencv-python>=4.7.0
Pillow>=9.4.0
numpy>=1.23.0
pandas>=1.5.0